Add a Ruby grammar to show off a bunch of the new features.
The fact that we can parse contextually lets us do a lot of stuff we couldn't before. For instance, it lets us highlight method parameters properly, because we can run the default-value capture back through the parser. It also lets us highlight a number of constructs that use {braces} without those rules stepping on one another.